Add files via upload
Fixed a few bugs in data_to_graph_v2; wrote an improved version data_to_graph_v3 that considered weight type in the loading phase. Added new pruning method prune_v2 to graph_utils that can guarantee connectivity not be impaired by the pruning process, but this version of pruning is computationally intensive. Added the new feature of writing a txt file representation of a pruned graph so that pruning only needs to be carried out once. The main block was written and seemed to work well.